Move Devices tests into a separate class.
Should have gone in with commit ca9b27d9b8cf6cfd85de9bf7f84ee6345ffadfb8.
